updates from feedback, new editor file, stub tests
