+++ packages/editor/src/lib/TldrawEditor.tsx
@@ -1,6 +1,6 @@
 import { MigrationSequence, Store } from '@tldraw/store'
 import { TLStore, TLStoreSnapshot } from '@tldraw/tlschema'
-import { Required, annotateError, lns } from '@tldraw/utils'
+import { Required, annotateError } from '@tldraw/utils'
 import React, {
 	ReactNode,
 	memo,
@@ -11,7 +11,6 @@ import React, {
 	useSyncExternalStore,
 } from 'react'
 
-import { useValue } from '@tldraw/state-react'
 import classNames from 'classnames'
 import { version } from '../version'
 import { OptionalErrorBoundary } from './components/ErrorBoundary'
@@ -22,10 +21,9 @@ import { TLUser, createTLUser } from './config/createTLUser'
 import { TLAnyBindingUtilConstructor } from './config/defaultBindings'
 import { TLAnyShapeUtilConstructor } from './config/defaultShapes'
 import { Editor } from './editor/Editor'
-import { LicenseManager } from './editor/managers/LicenseManager'
 import { TLStateNodeConstructor } from './editor/tools/StateNode'
 import { TLCameraOptions } from './editor/types/misc-types'
-import { useCanvasEvents } from './hooks/useCanvasEvents'
+import { Watermark } from './editor/watermark/Watermark'
 import { ContainerProvider, useContainer } from './hooks/useContainer'
 import { useCursor } from './hooks/useCursor'
 import { useDarkMode } from './hooks/useDarkMode'
@@ -41,10 +39,8 @@ import { useLocalStore } from './hooks/useLocalStore'
 import { useRefState } from './hooks/useRefState'
 import { useZoomCss } from './hooks/useZoomCss'
 import { TldrawOptions } from './options'
-import { featureFlags } from './utils/debug-flags'
 import { stopEventPropagation } from './utils/dom'
 import { TLStoreWithStatus } from './utils/sync/StoreWithStatus'
-import { watermarkDesktopSvg } from './watermarks'
 
 /**
  * Props for the {@link tldraw#Tldraw} and {@link TldrawEditor} components, when passing in a
@@ -524,118 +520,3 @@ export function useOnMount(onMount?: TLOnMountHandler) {
 		if (editor) return onMountEvent?.(editor)
 	}, [editor, onMountEvent])
 }
-
-const WATERMARK_SRC = `url('data:image/svg+xml;utf8,${encodeURIComponent(watermarkDesktopSvg)}') center 100% / 100% no-repeat`
-
-const Watermark = React.memo(() => {
-	const events = useCanvasEvents()
-
-	const ref = useRef<HTMLAnchorElement>(null)
-	useLayoutEffect(() => {
-		// eslint-disable-next-line deprecation/deprecation
-		if (ref?.current) ref.current.style.webkitMask = WATERMARK_SRC
-	}, [])
-
-	const editor = useEditor()
-	const showWatermark = useValue(
-		'show watermark',
-		() => {
-			const { width } = editor.getViewportScreenBounds()
-			if (width < 760) return false
-			if (!featureFlags.enableLicensing.get()) return false
-			return editor.getLicenseState() === 'unlicensed'
-		},
-		[editor]
-	)
-	const isDebugMode = useValue('debug mode', () => editor.getInstanceState().isDebugMode, [editor])
-	const isMenuOpen = useValue('is menu open', () => editor.getIsMenuOpen(), [editor])
-
-	if (!showWatermark) return null
-
-	const className = LicenseManager.classname
-	const imageClassName = LicenseManager.classname + lns(`watermark${version.replace(/\./g, '')}`)
-
-	return (
-		<>
-			<style>
-				{`
-/* ------------------- SEE LICENSE -------------------
-The tldraw watermark is part of tldraw's license. It is shown for unlicensed
-users. By using this library, you agree to keep the watermark's behavior, 
-keeping it visible, unobscured, and available to user-interaction.
-
-To remove the watermark, please purchase a license at tldraw.dev.
-*/
-
-.${className} {
-	position: absolute;
-	bottom: var(--space-2);
-	right: calc(42px);
-	width: 96px;
-	height: 32px;
-	z-index: 2147483647 !important;
-	pointer-events: ${isMenuOpen ? 'none' : 'all'};
-	background-color: color-mix(in srgb, var(--color-background) 62%, transparent);
-	border-radius: 5px;
-	padding: 2px;
-	box-sizing: content-box;
-}
-
-.${className}[data-debug='true'] {
-	bottom: 46px;
-}
-
-.${className} > a {
-	position: absolute;
-	width: 96px;
-	height: 32px;
-	pointer-events: none;
-	cursor: inherit;
-	color: var(--color-text);
-	background-color: currentColor;
-	opacity: .28;
-}
-
-@media (hover: hover) {
-	.${className}:hover {
-		background-color: var(--color-background);
-		transition: background-color 0.2s ease-in-out;
-		transition-delay: 0.32s;
-	}
-	.${className}:hover > a {
-		animation: delayed_link 0.2s forwards ease-in-out;
-		animation-delay: 0.32s;
-	}
-}
-
-@keyframes delayed_link {
-	0% {
-		cursor: inherit;
-		opacity: .38;
-		pointer-events: none;
-	}
-	100% {
-		cursor: pointer;
-		opacity: 1;
-		pointer-events: all;
-	}
-}
-`}
-			</style>
-			<div className={className} data-debug={isDebugMode} draggable={false} {...events}>
-				<a
-					ref={ref}
-					className={imageClassName}
-					href="https://tldraw.dev"
-					target="_blank"
-					rel="noreferrer"
-					draggable={false}
-					onPointerDown={stopEventPropagation}
-					style={{
-						mask: WATERMARK_SRC,
-					}}
-				/>
-			</div>
-		</>
-	)
-})
